requirements.txt
cachetools==5.5.0
DateTime==5.5
google-auth==2.34.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
numpy==2.1.1
oauthlib==3.2.2
pandas==2.2.2
prettytable==3.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pytz==2024.2
requests-oauthlib==2.0.0
rsa==4.9
timedelta==2020.12.3
tzdata==2024.1
zope.interface==7.0.3
